chatbot-backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "fastapi-cors>=0.0.6",
    "google-generativeai>=0.8.6",
    "openai>=2.14.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
